Issue #1: Implement downloader for torrent files
The goal of this issue was to implement the base for the downloader. The
implementation was meant to be very simple. The only real feature
necessary is that the user could provide a download directory and a path
to a torrent file and the application should download the torrent to
that directory and close cleanly afterwards.

Important classes added with these changes:
- DownloadEngine
- DownloadReference
- DownloadManager
- EventListener