Issue #3: Show download progress to the user
Before, the user was not shown any progress information. He only knew
when the download was started and that download finished because once
the application closed. The point of this issue was to display progress
information to the user.

Now, at every second the user is updated with the status of the
download. Information currently shown includes:
- The name of the download
- The state
- The progress percentage
- The download rate

New introductions to the API:
- DownloadHandle
- DownloadState
- DownloadStatus

Changes to existing API:
- DownloadEngine's add() methods now return the handle for the new
  download