Issue #4: Provide way to stop downloads through code
The download is stopped through the handle, as opposed to removing it
through the engine by passing the handle. The latter option requires
the use of reflection. No good. In the future, it could be added the
option to remove a download from the engine by specifying the reference.
However, at this point, the download is stopped by calling the handle's
stop() method.

Changes to existing API:
- New 'stop()' method added to DownloadHandle
- New 'isValid' property added to DownloadHandle