Issue #13: Persist Tracker's state and metadata
The tracker records information about the TV shows available in the
TV show provider, as well as the episodes that has already found.
Before, that information was kept information because the
TrackerRepository and TVShowIDMapper implementations did not use any
kind of persisting storage, only memory only data structures.

Now, there are two new implementations which persist data into a
database. The database being used here is H2 because it can be embedded
with the application. However, this can be configured later to use any
other database. The Exposed framework is used to abstract the database
operations.

The path to the database file can be configured through the property
`datasource.savePath`.