Issue #12: Resume previous downloads on startup
The DownloaderService now keeps a download queue containing the
downloads added to the DownloadEngine. This queue is persisted, allowing
the service to re-add to the DownloadEngine every download stored in the
queue when the application starts up.

The DownloaderEngine provides now a mechanism to the inform other
components that a download has finished downloading. This feature is
used by the DownloaderService to remove downloads from the DownloadQueue
once they have finished.