Issue #17: Avoid episode being found but bot downloaded due to crash
Before, if an error occurs while the downloader service is being
notified, then it was possible tha the episode was marked as found but
not included in the downloading queue. Consequently, that episode would
never be downloaded.

That could occur because the episode was marked as found before
notifying the downloader service. That is no longer the case. The order
of these steps was reversed to avoid the mentioned corner case.