Issue #22: Remove downloads of episodes corresponding to a TV show that was untracked
The `DownloaderService` now provides a method to remove the downloads
corresponding to a TV show. The `TrackerService` now calls that method
when a TV show is removed from the tracking list.

The `DownloadQueue` is now called `EpisodeDownloadQueue` and it holds
not only the download references and links, but also the IDs of the TV
shows to allow these to be removed.