Issue #38: Generate deterministic IDs for TV shows
Random TV show IDs were being generated and these random IDs were being
used by the application instead of the provider's IDs. That was the case
to abstract the TV show ID from the actual provider. However, it doesn't
make sense for the generated ID for the same TV Show to be different
across different executions of the application.

This commit addresses that issue by ensuring the ID generated for a
given TV show is always the same across different executions of the
application.

At this point we are using the ID directly from the provider. That may
change in the future if we decide to change providers or support
different ones.

commit a59a95908495b2e0ac3c693c9c9d1aec7c708d41
Author: David Fialho <dfialho@protonmail.com>
Date:   Thu Oct 11 16:37:05 2018 +0100

    Issue 38: Use TV show provider's ID directly

commit 334e12f30e1fbdf259614a192a1ac37b043cadf1
Author: David Fialho <dfialho@protonmail.com>
Date:   Thu Oct 11 16:26:37 2018 +0100

    Issue 38: Turn TV show ID into a string