Issue #56: Fix dependency conflicts with kotlin version
When running the build some warnings were output indicating a conflict
between kotlin versions (1.3.31 and 1.2.31). The old version was coming
as an intrinsic dependency from `kotlintest`. Bumping the version of
`kotlintest` to `3.3.2` solved the issue.

Additionally, the ktor version was bumped to 1.1.4 which already
includes kotlin 1.3.30. The project's kotlin version was also updated
to align with the one used by ktor.